+++ examples/browser-sync/index.js
@@ -1,13 +1,13 @@
 /**
  * Module dependencies.
  */
-var browserSync = require('browser-sync').create();
-var proxy = require('../_proxy'); // require('http-proxy-middleware');
+const browserSync = require('browser-sync').create();
+const { createProxyMiddleware } = require('../../dist'); // require('http-proxy-middleware');
 
 /**
  * Configure proxy middleware
  */
-var jsonPlaceholderProxy = proxy('/users', {
+const jsonPlaceholderProxy = createProxyMiddleware('/users', {
   target: 'http://jsonplaceholder.typicode.com',
   changeOrigin: true, // for vhosted sites, changes host header to match to target's host
   logLevel: 'debug'